requirements.txt
numpy==1.11.0
multiprocess==0.70.5
pysam==0.11.2.2
regex==2017.9.23
numba==0.35.0
pandas==0.19.1
scipy==0.18.1
matplotlib==1.5.1
PyYAML==3.12
